Backend/requirements.txt
aniso8601==9.0.1
arrow==1.2.3
attrs==22.2.0
aws-lambda-builders==1.30.0
aws-sam-cli==1.82.0
aws-sam-translator==1.65.0
aws-wsgi==0.2.7
bcrypt==4.0.1
beautifulsoup4==4.12.2
binaryornot==0.4.4
boto3==1.26.126
botocore==1.29.126
certifi==2022.12.7
cffi==1.15.1
cfn-lint==0.77.5
chardet==5.1.0
charset-normalizer==3.1.0
chevron==0.14.0
click==8.1.3
colorama==0.4.4
cookiecutter==2.1.1
cryptography==39.0.2
dateparser==1.1.8
defusedxml==0.7.1
dnspython==2.3.0
docker==4.2.2
docutils==0.16
email-validator==1.3.1
exceptiongroup==1.1.1
Flask==2.0.3
Flask-Bcrypt==1.0.1
Flask-HTTPAuth==4.7.0
flask-mongoengine==1.0.0
Flask-RESTful==0.3.9
Flask-WTF==1.1.1
httpie==3.2.1
idna==3.4
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jinja2-time==0.2.0
jmespath==0.10.0
jschema-to-python==1.2.3
jsonpatch==1.32
jsonpickle==3.0.1
jsonpointer==2.3
jsonschema==4.17.3
junit-xml==1.9
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
mongoengine==0.27.0
moto==4.1.8
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
packaging==23.0
pbr==5.11.1
pluggy==1.0.0
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.7
Pygments==2.15.1
pymongo==4.3.3
pyOpenSSL==23.0.0
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.2.2
pytest-mock==3.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==8.0.1
pytz==2023.3
PyYAML==5.4.1
regex==2023.3.23
requests==2.28.2
requests-mock==1.10.0
requests-toolbelt==0.10.1
responses==0.23.1
rich==13.3.4
rsa==4.7.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
sarif-om==1.0.4
serverlessrepo==0.1.10
six==1.16.0
soupsieve==2.4
sympy==1.12
text-unidecode==1.3
tomli==2.0.1
tomlkit==0.11.7
types-PyYAML==6.0.12.9
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==3.0
urllib3==1.26.15
watchdog==2.1.2
websocket-client==1.5.1
Werkzeug==2.2.3
WTForms==3.0.1
xmltodict==0.13.0
